requirements.txt
# eBay Research Tool Dependencies

# Web Automation
playwright==1.40.0

# Data Processing
pandas==2.1.1
openpyxl==3.1.2  # Excel support for pandas

# Database
sqlalchemy==2.0.23

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Security
cryptography==41.0.5
python-dotenv==1.0.0

# CLI
typer==0.9.0
rich==13.6.0  # Rich text and formatting in terminal

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3

# Utilities
tenacity==8.2.3  # Retrying library
tqdm==4.66.1  # Progress bars
